+++ apps/bot/src/commands/fun/fun.ts
@@ -19,7 +19,7 @@ import fetch from "node-fetch";
 import { CommandInteraction } from "discord.js";
 import { ApplicationCommandRegistry, Command } from "@sapphire/framework";
 import { ApplicationCommandType } from "discord-api-types";
-import { getGuildIDS } from "../utils/parsers/envUtils";
+import { getGuildIDS } from "../../utils/parsers/envUtils";
 
 export class Fun extends Command {
   public override async chatInputRun(interaction: CommandInteraction) {
